+++ packages/headless/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sit-onyx/headless
 
+## 1.0.0-beta.1
+
+### Major Changes
+
+- 2e29c45: - BREAKING CHANGE: `createTooltip` now provides a `root` element, which is used to listen for outside clicks
+  - new `useGlobalListener` composable helper
+
 ## 1.0.0-beta.0
 
 ### Major Changes